# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Company Steps Leaderboard API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /teams/{teamID}/users/{userID}/counters/increment:
      post:
        summary: Increment the step counter of a user in a team by a value
        operationId: incrementCounterByValue
        parameters:
          - name: teamID
            in: path
            required: true
            schema:
              type: string
          - name: userID
            in: path
            required: true
            schema:
              type: string
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  incrementValue:
                    type: integer
                    description: The value to increment the step counter by.
        responses:
          '200':
            description: OK
          '404':
            description: User not found
...